+++ Libmathft/vectors_and_matrices/v_norm.cu
@@ -6,7 +6,7 @@
 /*   By: tgros <tgros@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/02/21 13:16:25 by jwalsh            #+#    #+#             */
-/*   Updated: 2017/03/31 15:26:14 by tgros            ###   ########.fr       */
+/*   Updated: 2017/04/26 11:17:11 by tgros            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@
 __host__ __device__
 t_vec3	v_norm(t_vec3 v)
 {
-	double	len;
+	float	len;
 
 	len = v_length(v);
 	if (len > 0)
